+++ grpc/grpc-client-extension/src/main/resources/META-INF/services/jakarta.enterprise.inject.spi.Extension
@@ -1,0 +1,1 @@
+hu.icellmobilsoft.sampler.grpc.client.extension.GrpcClientExtension
