+++ grpc/grpc-server-extension/pom.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>hu.icellmobilsoft.sampler.grpc</groupId>
+        <artifactId>grpc</artifactId>
+        <version>2.0.0-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>grpc-server-extension</artifactId>
+
+    <dependencies>
+        <dependency>
+            <groupId>hu.icellmobilsoft.coffee</groupId>
+            <artifactId>coffee-tool</artifactId>
+        </dependency>
+       <!--TODO BaseException kozpontositas-->
+       <dependency>
+            <groupId>hu.icellmobilsoft.sampler.dto</groupId>
+            <artifactId>dto</artifactId>
+        </dependency>
+		<dependency>
+			<groupId>hu.icellmobilsoft.sampler.grpc</groupId>
+			<artifactId>grpc-core</artifactId>
+		</dependency>
+        
+    </dependencies>
+
+</project>
