--- sample/sample-grpc-server-service/src/main/java/hu/icellmobilsoft/sampler/sample/grpc/server/IGrpcServerExecutor.java
@@ -1,30 +1,0 @@
-/*-
- * #%L
- * Sampler
- * %%
- * Copyright (C) 2022 - 2023 i-Cell Mobilsoft Zrt.
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-package hu.icellmobilsoft.sampler.sample.grpc.server;
-
-/**
- * Executor interface
- * 
- * @author czenczl
- * @version 2.0.0
- */
-public interface IGrpcServerExecutor extends Runnable {
-
-}
