--- sample/sample-grpc-server-service/src/main/resources/META-INF/services/jakarta.enterprise.inject.spi.Extension
@@ -1,1 +1,0 @@
-hu.icellmobilsoft.sampler.sample.grpc.client.extension.GrpcClientExtension
